Fixing the "Freezing" Issue When Viewing the Log (server_log.txt) in This Code
The issue of the program freezing when clicking to view the log was caused by how the log file was being loaded. Previously, the file was read synchronously (in the main thread, which is also the UI thread). The solution involves changing the file reading method to asynchronous (reading the file in a background thread and then updating the UI in the main thread), as follows:

Cause of the freeze:

Initially, when clicking to view the log, the program read the entire server_log.txt file in the main thread (UI thread).

If the log file is large or disk IO is slow, the UI becomes unresponsive (freezes) until the file reading is completed.

Solution:

Use threading.Thread to read the log file in a background thread (so it doesn't block the UI).

Once the file is read, use self.root.after(0, update_textbox) to safely update the text in the textbox from the main thread.

Result:

The UI no longer freezes or becomes unresponsive while loading the log file, no matter how large the file is.